refactor(dashboard): Migrate to modular, folder-based routing
